+++ src/Symfony/Component/Translation/Dumper/XliffFileDumper.php
@@ -21,6 +21,11 @@ use Symfony\Component\Translation\MessageCatalogue;
  */
 class XliffFileDumper extends FileDumper
 {
+    public function __construct(
+        private string $extension = 'xlf',
+    ) {
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -52,7 +57,7 @@ class XliffFileDumper extends FileDumper
      */
     protected function getExtension(): string
     {
-        return 'xlf';
+        return $this->extension;
     }
 
     private function dumpXliff1(string $defaultLocale, MessageCatalogue $messages, ?string $domain, array $options = [])